Refactoring scraper executor func
