+++ pcts_scrapers_api/pcts_scrapers_api/celery.py
@@ -55,7 +55,7 @@ def setup_periodic_scrapers(sender: Celery, **kwargs):
     sender.add_periodic_task(
         crontab(minute='0', hour='4', day_of_week='*',
                 day_of_month='*', month_of_year='*'),
-        tasks.incra_scraper_group.subtask(keywords=KEYWORDS),
+        tasks.incra_scraper_group.subtask(kwargs={"keywords": KEYWORDS}),
         name="incra_scraper_group",
     )
 
@@ -63,6 +63,6 @@ def setup_periodic_scrapers(sender: Celery, **kwargs):
     sender.add_periodic_task(
         crontab(minute='0', hour='5', day_of_week='*',
                 day_of_month='*', month_of_year='*'),
-        tasks.mpf_scraper.subtask(keywords=KEYWORDS),
-        name="mpf_scraper",
+        tasks.mpf_scraper_group.subtask(kwargs={"keywords": KEYWORDS}),
+        name="mpf_scraper_group",
     )
