+++ scripts/start_scrapers.sh
@@ -3,4 +3,4 @@
 export PCTS_DOCUMENTS_API_URL=http://127.0.0.1:8000
 export PCTS_DOCUMENTS_API_RECORDS_ENDPOINT=documents/documents/
 
-python run_scrapers.py
+python scraper_executor.py
